requirements.txt
torch==2.4.1
transformers==4.38.2  # Updated to a version compatible with sentence-transformers==3.1.1
deepspeech==0.9.3
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==3.1.1
sounddevice==0.4.6
scipy==1.10.1